requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.4.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==21.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==4.0.2  # https://github.com/andymccurdy/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.46.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.10.0 # https://github.com/adamchainz/django-cors-headers

# External libraries
# ------------------------------------------------------------------------------
#youtube-dl
yt-dlp>=2023.07.06
gallery-dl>=1.25.8
PyExifTool
psutil
pygal
cairosvg
pillow
python-magic
telethon
cryptg
aiohttp
opencv-python-headless

# Indexing
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil

# Utils
django-taggit==2.1.0
django-markdownify==0.9.0
django-tinymce==3.4.0
django-q==1.3.6
django-minio-storage==0.3.10
django-notifications-hq==1.6.0
